updated for changed to lcfgutils_string_djbhash function
git-svn-id: https://svn.lcfg.org/svn/source/trunk/lcfg-core@32891 c6336a39-18d7-4234-9af0-a2bff563b40a

+++ include/tags.h
@@ -41,9 +41,9 @@
 
 struct LCFGTag {
   /*@{*/
-  char * name;       /**< The tag name */
-  size_t name_len;   /**< The length of the tag name */
-  unsigned int hash; /**< The hash value for the tag name */
+  char * name;        /**< The tag name */
+  size_t name_len;    /**< The length of the tag name */
+  unsigned long hash; /**< The hash value for the tag name */
   /*@}*/
   unsigned int _refcount;
 };
